Pre-commit audit: fix 5 bugs found in bilateral issues system
1. nudgeIssueRelations: Supabase JS client returns {error}, does not
   throw — changed try/catch to check .error property on response
2. processArbitration idempotency: was checking for ANY status_changed
   event which would false-positive on partial transitions — now checks
   specifically for arbitration_ruling metadata
3. spawnIncidentFromIssue stat mapping: Intl_Reputation_b mapped to
   intl_reputation instead of international_reputation via string
   slicing — replaced fragile key parsing with direct stat calls
4. Tension column: was INTEGER, losing fractional drift (0.25/tick from
   competitive modifiers) every tick — changed to NUMERIC(4,2)
5. Modal AP cost: diplomatic friction +1 AP penalty wasn't shown in
   action confirmation modal — now passes modifiers and applies check

+++ sql/migrations/20260405_bilateral_issues.sql
@@ -23,8 +23,8 @@ CREATE TABLE IF NOT EXISTS bilateral_issues (
     nation_a_id         UUID NOT NULL REFERENCES nations(id) ON DELETE CASCADE,
     nation_b_id         UUID NOT NULL REFERENCES nations(id) ON DELETE CASCADE,
 
-    -- Tension: 0-10 integer. Thresholds: Low 0-2, Moderate 3-5, High 6-8, Critical 9-10. At 10 → escalates to incident.
-    tension             INTEGER NOT NULL DEFAULT 0 CHECK (tension >= 0 AND tension <= 10),
+    -- Tension: 0-10 (quarter-step precision). Thresholds: Low 0-2, Moderate 3-5, High 6-8, Critical 9-10. At 10 → escalates to incident.
+    tension             NUMERIC(4,2) NOT NULL DEFAULT 0 CHECK (tension >= 0 AND tension <= 10),
 
     -- Favor: -5 to +5. Negative = favors nation_a, positive = favors nation_b. 0 = neutral.
     favor               NUMERIC(4,2) NOT NULL DEFAULT 0 CHECK (favor >= -5 AND favor <= 5),
